pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-content-factory"
version = "0.1.0"
description = "An AI-powered multi-agent content creation engine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Content Creation Engine Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "crewai>=0.28.0",
    "langchain>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
audio = [
    "audiocraft>=1.0.0",
    "torchaudio>=2.0.0",
]
full = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]

[project.scripts]
content-factory = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
